create SelectQueryDropdown.tsx
